+++ crates/nostr-relay-pool/src/prelude.rs
@@ -13,9 +13,10 @@ pub use async_utility::futures_util::StreamExt; // Needed for `RelayPool::stream
 pub use nostr::prelude::*;
 pub use nostr_database::*;
 
+// Internal modules
 pub use crate::pool::constants::*;
 pub use crate::pool::options::*;
-// Internal modules
 pub use crate::pool::{self, *};
 pub use crate::relay::{self, *};
+pub use crate::stream::*;
 pub use crate::*;